Refactor metadata command with tasks
