+++ src/docbuild/tasks/__init__.py
@@ -1,0 +1,1 @@
+"""Task packages for reusable docbuild execution logic."""
