+++ src/docbuild/tasks/repository/__init__.py
@@ -1,0 +1,6 @@
+"""Reusable repository update and worktree helpers for task modules."""
+
+from .sync import update_managed_repositories
+from .worktree import shared_worktrees
+
+__all__ = ["shared_worktrees", "update_managed_repositories"]
